[Refactor] Container에 존재하는 presentational 로직 추출
프레젠테이셔널은 컨테이너의 관심사가 아니기 때문에 이를 분리함
