+++ client/src/components/issueDetail/container/AssignedSidebarItem.jsx
@@ -2,37 +2,12 @@ import { React, useContext } from 'react';
 import { PropTypes } from 'prop-types';
 import { IssuesContext } from '../../../stores/IssueStore';
 
-const styles = {
-  defaultMessage: {
-    color: '#586069',
-    fontWeight: '100',
-  },
-  selfAssignButton: {
-    '&:hover': {
-      color: 'blue',
-    },
-  },
-  progress: {
-    height: '10px',
-    backgroundColor: '#4CAF50',
-  },
-  bar: {
-    width: '100%',
-    backgroundColor: '#ddd',
-  },
-  item: {
-    width: 'max-content',
-    padding: '2px 10px',
-    margin: '5px 5px 0 0',
-    boxSizing: 'border-box',
-    borderRadius: '15px',
-  },
-};
+import AssignedItems from '../presentational/AssignedItems';
+import AssignedItemsWithMilestone from '../presentational/AssignedItemsWithMilestone';
 
-export default function AssignedSidebarItem({
-  title, assigned,
-}) {
+export default function AssignedSidebarItem({ title, assigned }) {
   const { issues } = useContext(IssuesContext);
+
   const progressPercentage = (milestoneId) => {
     let closedCount = 0;
     const checkPoints = issues.filter((checkpoint) => checkpoint.milestoneId === +milestoneId);
@@ -41,22 +16,11 @@ export default function AssignedSidebarItem({
     });
     return checkPoints.length ? (closedCount * 100) / checkPoints.length : 0;
   };
+
   return (
     title === 'Milestone'
-      ? assigned.map((element) => (
-        <div css={{ ...styles.item, width: '200px' }}>
-          <div css={styles.bar}>
-            <div css={{ ...styles.progress, width: `${progressPercentage(element.id)}%` }} />
-          </div>
-          {element.title}
-        </div>
-      ))
-      : assigned.map((element) => (
-        <div css={{ ...styles.item, background: element.color, display: element.name ? 'inline-block' : 'block' }}>
-          {element.name || element.title || element.nickname}
-        </div>
-      ))
-  );
+      ? <AssignedItemsWithMilestone assigned={assigned} progressPercentage={progressPercentage} />
+      : <AssignedItems assigned={assigned} />);
 }
 
 AssignedSidebarItem.propTypes = {
